Module lock_keeper_client::api

Full implementation of the public API for the Lock Keeper client library.

This API is designed for use with a local client application - that is, an application running directly on the device of an asset owner. The inputs the asset owner provides should be passed directly to this API without being sent to a separate machine.

Modules

arbitrary_secrets

Functions

create_digital_asset_key

Generate a new, distributed digital asset key with the given use parameters for the [UserId], and compatible with the specified blockchain.

request_transaction_signature

Request a signature on a transaction from the key server.

retrieve_audit_event_log

Retrieve the log of audit events from the key server for a specified asset owner; optionally, filter for audit events associated with the specified [KeyId].

retrieve_public_key_by_id

Retrieve the public key info for the specified key associated with the user.

retrieve_public_keys

Retrieve the public key info for all keys associated with the specified user that are stored at the key server.

set_user_key_policy

Set an asset-owner-specified key policy for a delegated key.