Function lock_keeper_client::api::request_transaction_signature

pub fn request_transaction_signature(
    transaction_approval_request: TransactionApprovalRequest
) -> Result<TransactionSignature, LockKeeperClientError>

Request a signature on a transaction from the key server.

Among the parameters in the TransactionApprovalRequest, the KeyId must correspond to a key owned by the UserId, and the UserId must match the user authenticated in the crate::LockKeeperClient.

Assumption: A TransactionApprovalRequest originates either with the asset owner or a key fiduciary. This is cryptographically enforced with an authenticated crate::LockKeeperClient between the key server and one of the asset owner or a key fiduciary. This request will fail if the calling party is not from one of those entities.

Output: If successful, returns a TransactionSignature as specified in the original TransactionApprovalRequest – that is, over the Transaction, and using the key corresponding to the KeyId.