Struct lock_keeper_client::client::LockKeeperClient

pub struct LockKeeperClient {
    session: Session,
    config: Config,
    account_name: AccountName,
    user_id: UserId,
    master_key: MasterKey,
    tonic_client: LockKeeperRpcClient<Client<HttpsConnector<HttpConnector>, UnsyncBoxBody<Bytes, Status>>>,
    pub(crate) rng: Arc<Mutex<StdRng>>,
}

A LockKeeperClient is an abstraction over client operations; that is, it wraps around the state and infrastructure necessary to make requests to the key server. It handles confidentiality, integrity, and authentication of communication with the server. A LockKeeperClient exists for the duration of one authenticated session, during which multiple requests can be made to the server.

TODO #30: This abstraction needs a lot of design attention.

Fields
session: Sessionconfig: Configaccount_name: AccountNameuser_id: UserIdmaster_key: MasterKeytonic_client: LockKeeperRpcClient<Client<HttpsConnector<HttpConnector>, UnsyncBoxBody<Bytes, Status>>>rng: Arc<Mutex<StdRng>>
Implementations

impl LockKeeperClient

pub(crate) async fn handle_authentication(
channel: ClientChannel<Unauthenticated>,
rng: Arc<Mutex<StdRng>>,
account_name: &AccountName,
password: &Password
) -> Result<AuthenticateResult, LockKeeperClientError>

impl LockKeeperClient

pub(crate) async fn handle_create_storage_key<T: CryptoRng + RngCore>(
&self,
channel: ClientChannel<Authenticated<StdRng>>,
rng: Arc<Mutex<T>>,
master_key: MasterKey
) -> Result<(), LockKeeperClientError>

Creates a storage key and sends it to the key server

impl LockKeeperClient

pub(crate) async fn handle_generate_secret(
&self,
channel: ClientChannel<Authenticated<StdRng>>,
request_id: Uuid
) -> Result<GenerateResult, LockKeeperClientError>

impl LockKeeperClient

pub(crate) async fn handle_get_user_id(
channel: ClientChannel<Authenticated<StdRng>>
) -> Result<UserId, LockKeeperClientError>

impl LockKeeperClient

pub(crate) async fn handle_import_signing_key(
&self,
channel: ClientChannel<Authenticated<StdRng>>,
key_material: Import
) -> Result<KeyId, LockKeeperClientError>

impl LockKeeperClient

pub(crate) async fn handle_registration(
channel: ClientChannel<Unauthenticated>,
rng: Arc<Mutex<StdRng>>,
account_name: &AccountName,
password: &Password
) -> Result<MasterKey, LockKeeperClientError>

impl LockKeeperClient

pub(crate) async fn handle_remote_generate_signing_key(
&self,
channel: ClientChannel<Authenticated<StdRng>>
) -> Result<RemoteGenerateResult, LockKeeperClientError>

impl LockKeeperClient

pub(crate) async fn handle_remote_sign_bytes(
&self,
channel: ClientChannel<Authenticated<StdRng>>,
key_id: KeyId,
bytes: impl Signable
) -> Result<Signature, LockKeeperClientError>

impl LockKeeperClient

pub(crate) async fn handle_retrieve_secret(
&self,
channel: ClientChannel<Authenticated<StdRng>>,
key_id: &KeyId,
context: RetrieveContext,
request_id: Uuid
) -> Result<Option<LocalStorage<Secret>>, LockKeeperClientError>

Handles the retrieval of arbitrary secrets ([lock_keeper::crypto::Secret]) only.

pub(crate) async fn handle_retrieve_signing_key(
&self,
channel: ClientChannel<Authenticated<StdRng>>,
key_id: &KeyId,
context: RetrieveContext
) -> Result<Option<LocalStorage<SigningKeyPair>>, LockKeeperClientError>

Handles the retrieval of signing keys ([lock_keeper::crypto::SigningKeyPair]) only.

impl LockKeeperClient

pub(crate) async fn handle_retrieve_audit_events(
&self,
channel: ClientChannel<Authenticated<StdRng>>,
event_type: EventType,
options: AuditEventOptions
) -> Result<Vec<AuditEvent>, LockKeeperClientError>

impl LockKeeperClient

pub async fn health(config: &Config) -> Result<(), LockKeeperClientError>

Ping the server to make sure it is running and reachable

pub async fn logout(&self) -> LockKeeperResponse<()>

Expire the current session and session key for this user.

pub async fn authenticated_client(
account_name: &AccountName,
password: &Password,
config: &Config
) -> LockKeeperResponse<Self>

Authenticate to the Lock Keeper key server as a previously registered user.

Output: If successful, returns a LockKeeperClient.

pub async fn register(
account_name: &AccountName,
password: &Password,
config: &Config
) -> LockKeeperResponse<()>

Register a new user who has not yet interacted with the service.

This only needs to be called once per user; future sessions can be created with LockKeeperClient::authenticated_client().

Output: Returns Ok if successful. To perform further operations, use Self::authenticated_client().

async fn register_helper(
account_name: &AccountName,
password: &Password,
config: &Config,
request_id: Uuid
) -> Result<(), LockKeeperClientError>

pub async fn export_secret(&self, key_id: &KeyId) -> LockKeeperResponse<Export>

Export an arbitrary key from the key servers.

Calling this function on a signing key will generate an error. Output: If successful, returns the requested key material in byte form.

async fn export_secret_helper(
&self,
key_id: &KeyId,
request_id: Uuid
) -> Result<Export, LockKeeperClientError>

pub async fn export_signing_key(
&self,
key_id: &KeyId
) -> LockKeeperResponse<Export>

Export signing key pair material from the key servers.

Calling this function on an arbitrary key will generated an error. Output: If successful, returns the requested key material in byte form.

async fn export_signing_key_helper(
&self,
key_id: &KeyId,
request_id: Uuid
) -> Result<Export, LockKeeperClientError>

pub async fn generate_secret(&self) -> LockKeeperResponse<GenerateResult>

Generate an arbitrary secret client-side, store this secret in the key server.

async fn generate_secret_helper(
&self,
request_id: Uuid
) -> Result<GenerateResult, LockKeeperClientError>

pub async fn import_signing_key(
&self,
key_material: Import
) -> LockKeeperResponse<KeyId>

Import signing key material to the key server

async fn import_signing_key_helper(
&self,
key_material: Import,
request_id: Uuid
) -> Result<KeyId, LockKeeperClientError>

pub async fn retrieve_secret(
&self,
key_id: &KeyId,
context: RetrieveContext
) -> LockKeeperResponse<Option<LocalStorage<Secret>>>

Retrieve a secret from the key server by [KeyId]

This operation will fail if it is called on a signing key.

async fn retrieve_secret_helper(
&self,
key_id: &KeyId,
context: RetrieveContext,
request_id: Uuid
) -> Result<Option<LocalStorage<Secret>>, LockKeeperClientError>

pub async fn remote_generate(&self) -> LockKeeperResponse<RemoteGenerateResult>

Request that the server generate a new signing key.

async fn remote_generate_helper(
&self,
request_id: Uuid
) -> Result<RemoteGenerateResult, LockKeeperClientError>

pub async fn remote_sign_bytes(
&self,
key_id: KeyId,
bytes: impl Signable
) -> LockKeeperResponse<Signature>

Sign an arbitrary blob of bytes with a remotely generated [SigningKeyPair][lock_keeper::crypto::SigningKeyPair] and return the resulting [Signature].

async fn remote_sign_bytes_helper(
&self,
key_id: KeyId,
bytes: impl Signable,
request_id: Uuid
) -> Result<Signature, LockKeeperClientError>

pub async fn retrieve_audit_event_log(
&self,
event_type: EventType,
options: AuditEventOptions
) -> LockKeeperResponse<Vec<AuditEvent>>

Retrieve the log of audit events from the key server for the authenticated asset owner; optionally, filter for audit events associated with the specified [KeyId].

The log of audit events includes context about any action requested and/or taken on the digital asset key, including which action was requested and by whom, the date, details about approval or rejection from each key server, the policy engine, and each asset fiduciary (if relevant), and any other relevant details.

The [lock_keeper::types::database::user::UserId] must match the asset owner authenticated in the crate::LockKeeperClient, and if specified, the [KeyId] must correspond to a key owned by the [lock_keeper::types::database::user::UserId].

Output: if successful, returns a String representation of the logs.

async fn retrieve_audit_event_log_helper(
&self,
event_type: EventType,
options: AuditEventOptions,
request_id: Uuid
) -> Result<Vec<AuditEvent>, LockKeeperClientError>

impl LockKeeperClient

pub fn user_id(&self) -> &UserId

Get [UserId] for the authenticated client.

pub fn account_name(&self) -> &AccountName

Get [AccountName] for the authenticated client.

pub fn session_key(&self) -> &OpaqueSessionKey

Get [OpaqueSessionKey] for the authenticated client.

pub(crate) fn tonic_client(
&self
) -> LockKeeperRpcClient<Client<HttpsConnector<HttpConnector>, UnsyncBoxBody<Bytes, Status>>>

pub(crate) async fn connect(
config: &Config
) -> Result<LockKeeperRpcClient<Client<HttpsConnector<HttpConnector>, UnsyncBoxBody<Bytes, Status>>>>

Create a tonic client object and return it to the client app.

The returned client should be stored as part of the LockKeeperClient state.

pub(crate) async fn authenticate(
client: Option<LockKeeperRpcClient<Client<HttpsConnector<HttpConnector>, UnsyncBoxBody<Bytes, Status>>>>,
account_name: &AccountName,
password: &Password,
config: &Config,
request_id: Uuid
) -> Result<Self>

pub(crate) fn create_metadata(
&self,
action: ClientAction,
request_id: Uuid
) -> RequestMetadata

pub(crate) async fn create_channel(
client: &mut LockKeeperRpcClient<Client<HttpsConnector<HttpConnector>, UnsyncBoxBody<Bytes, Status>>>,
metadata: &RequestMetadata
) -> Result<ClientChannel<Unauthenticated>>

Helper to create the appropriate [ClientChannel] to send to tonic handler functions based on the client’s action. This function creates unauthenticated channels. Use create_authenticated_channel to create an authenticated channel.

This function will return an error if the [ClientAction] requires authentication.

pub(crate) async fn create_authenticated_channel(
client: &mut LockKeeperRpcClient<Client<HttpsConnector<HttpConnector>, UnsyncBoxBody<Bytes, Status>>>,
metadata: &RequestMetadata,
session_key: OpaqueSessionKey,
rng: Arc<Mutex<StdRng>>
) -> Result<ClientChannel<Authenticated<StdRng>>>

Helper to create the appropriate [ClientChannel] to send to tonic handler functions based on the client’s action. This function creates authenticated channels. Use create_channel to create an unauthenticated channel.

This function will return an error if the [ClientAction] requires an unauthenticated channel.

pub(crate) async fn handle_logout(&self, request_id: Uuid) -> Result<()>

Close a session.

Outputs: None, if successful.

pub(crate) async fn retrieve_storage_key(
&self,
request_id: Uuid
) -> Result<StorageKey>

Retrieve the [lock_keeper::crypto::Encrypted<StorageKey>] that belongs to the user specified by user_id