Module lock_keeper_client::api

Full implementation of the public API for the Lock Keeper client library.

This API is designed for use with a local client application - that is, an application running directly on the device of an asset owner. The inputs the asset owner provides should be passed directly to this API without being sent to a separate machine.

Modules

authenticate ğŸ”’

create_storage_key ğŸ”’

generate ğŸ”’

import ğŸ”’

register ğŸ”’

remote_generate ğŸ”’

remote_sign_bytes ğŸ”’

retrieve ğŸ”’

retrieve_audit_events ğŸ”’

Structs

LocalStorage

Wrapper for secrets prepared for local storage

RemoteGenerateResult